requirements.txt
pandas
numpy
requests
beautifulsoup4
sentence-transformers
lightgbm
scikit-learn
streamlit>=1.28.0
plotly
konlpy
deep-translator
altair<5
pyyaml
